requirements.txt
llm
llm-gemini
fastapi
uvicorn[standard]
websockets
PyYAML
